+++ src/types.ts
@@ -12,8 +12,8 @@ export declare type Context = {
 	 * hash without leading '#'
 	 */
 	hash: string
-	redirect(path: string): 'stop'
-	navigate(path: string): void
+	redirect(path: string, includeQuery?: boolean): 'stop'
+	navigate(path: string, includeQuery?: boolean): void
 	back(): void
 	forward(): void
 	canGoBack(): boolean
@@ -103,4 +103,9 @@ export declare interface Settings {
 	 * @summary Set to false if you don't want to intercept links by default.
 	 */
 	interceptLinks: boolean
+	/**
+	 * @default true
+	 * @summaryt Set to false to make it so search query is cleared when navigating
+	 */
+	includeQuery: boolean
 }